Updated "install.sh" scripts to work with CI.

+++ telegram_bot/install.sh
@@ -6,6 +6,7 @@
 TELEGRAM_CONFIGURATOR_DEPLOY_ARGS=null;
 SKIP_DEPLOY=true;
 FORCE_BUILD=false;
+RUN=true;
 
 # Exits the installation process, but prints a message to command line before doing so.
 # :param $1: Colour of the output line. This will be reset before exiting.
@@ -69,13 +70,19 @@ if [ "$TELEGRAM_CONFIGURATOR_DEPLOY_ARGS" != "null" ] && [ "$SKIP_DEPLOY" == "tr
         TELEGRAM_CONFIGURATOR_DEPLOY_ARGS="--with-tests";
 fi
 
+if [ "$RUN" == "false" ]; then
+    message -1 "[INFO] The Telegram Configurator will not be executed."
+fi
 
 if ([ "$SKIP_DEPLOY" != "true" ] && [ "$SKIP_DEPLOY" != "false" ]) ||
-   ([ "$FORCE_BUILD" != "true" ] && [ "$FORCE_BUILD" != "false" ]); then
-     exit_with_message 1 "> usage: install.sh [FORCE_BUILD=true] [SKIP_DEPLOY=false] [TELEGRAM_CONFIGURATOR_DEPLOY_ARGS=<args>]
+   ([ "$FORCE_BUILD" != "true" ] && [ "$FORCE_BUILD" != "false" ]) ||
+   ([ "$RUN" != "true" ] && [ "$RUN" != "false" ]); then
+     exit_with_message 1 "> usage: install.sh [FORCE_BUILD=true] [SKIP_DEPLOY=false] [RUN=false]
+                            [TELEGRAM_CONFIGURATOR_DEPLOY_ARGS=<args>]
                          \n\t- FORCE_BUILD: builds the Telegram Configurator image even if it already exists. Defaults
                                to \"false\".
                          \n\t- SKIP_DEPLOY: omits all deploy steps. Defaults to \"true\".
+                         \n\t- RUN: executes the Configurator after building it. Defaults to \"true\".
                          \n\t- TELEGRAM_CONFIGURATOR_DEPLOY_ARGS: enables \"Expert Mode\", allowing to pass custom
                                args to the deploy script. Defaults to \"--with-tests\".
                          \nIMPORTANT: TELEGRAM_CONFIGURATOR_DEPLOY_ARGS must be used in conjunction with
@@ -93,20 +100,21 @@ message 4 "[COMPONENT] Building the Telegram Configurator component.";
 if [ "$(docker image ls | grep telegram_bot | wc -c | tr -d ' ')" == 0 ] || [ "$FORCE_BUILD" == "true" ]; then
     docker-compose build --build-arg DEPLOY_ARGS="${TELEGRAM_CONFIGURATOR_DEPLOY_ARGS}" telegram_bot;
     if [ $? != 0 ]; then
-        exit_with_message 1 "> The Telegram Configurator image could not be built." $?;
+        exit_with_message 1 "> The Telegram Configurator image could not be built." 1;
     fi
     else message -1 "[INFO] Running pre-built Telegram Configurator Docker image.";
 fi
 
 # Running the Telegram Configurator component
-message -1 "[INFO] Running the Telegram Configurator component.";
-echo "";
-docker run --rm -i -t --name telegram_bot diegohermida/telegram_bot:1.0;
-
+if [ "$RUN" == "true" ]; then
+    message -1 "[INFO] Running the Telegram Configurator component.";
+    echo "";
+    docker run --rm -i -t --name telegram_bot diegohermida/telegram_bot:1.0;
+fi
 
 # Displaying installation summary
 if [ $? != 0 ]; then
-        exit_with_message 1 "\n[ERROR] The Telegram Configurator did not exit normally. You should rerun this installer." $?;
+        exit_with_message 1 "\n[ERROR] The Telegram Configurator did not exit normally. You should rerun this installer." 1;
    else exit_with_message 2 "\n[SUCCESS] The Telegram Configurator was successful." 0;
 fi
 echo ""
